Create list of items (#5)
* feat: use databinding

* feat: navigate to main activity

* feat: add view model

* fix: bind my list fragment

* feat: create storage repository

* feat: add methods to storage repo

* feat: create list item

* fea

* feat: push list to adapter

* feat: display list

* feat: displaygit add .

* feat: use values

* feat: display images

+++ app/src/main/java/dog/snow/androidrecruittest/data/model/raw/RawAlbum.kt
@@ -19,10 +19,10 @@ data class RawAlbum(
     var id: Long = 0,
     @Unique
     @JsonProperty("id")
-    @Convert(converter = Converters.UID::class, dbType = Int::class)
+    @Convert(converter = Converters.UID::class, dbType = Long::class)
     val uId: UId,
     @JsonProperty("userId")
-    @Convert(converter = Converters.UID::class, dbType = Int::class)
+    @Convert(converter = Converters.UID::class, dbType = Long::class)
     val userUId: UId,
     @JsonProperty("title")
     @Convert(converter = Converters.TITLE::class, dbType = String::class)
